requirements.txt
requests==2.26.0
Django~=3.2
djangorestframework~=3.12.4
PyJWT==2.1.0
pytest~=6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3

django-filter~=22.1